Move the methods from IssueUtil to Backports.
